Conflicts:
	src/components/application_manager/include/application_manager/application_manager_impl.h
	src/components/transport_manager/src/transport_manager_impl.cc

+++ test/components/json_handler/src/formatter_test_helper.cc
@@ -66,7 +66,7 @@ void CFormatterTestHelper::compareObjects(
   if (SmartType_Array == first.getType()) {
     ASSERT_EQ(SmartType_Array, second.getType());
 
-    for (int i = 0; i < first.length(); i++) {
+    for (size_t i = 0; i < first.length(); i++) {
       compareObjects(first.getElement(i), second.getElement(i));
     }
   } else if (SmartType_Map == first.getType()) {
